pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mysql-analyzer-mcp"
version = "1.0.0"
description = "MySQL Database Analyzer MCP Server with CamelCase tables and snake_case columns"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MySQL Analyzer Team"}
]
keywords = ["mysql", "database", "analyzer", "mcp", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "mysql-connector-python>=8.0.32",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0", 
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/mysql-analyzer-mcp"
Repository = "https://github.com/your-org/mysql-analyzer-mcp"
Documentation = "https://github.com/your-org/mysql-analyzer-mcp#readme"
Issues = "https://github.com/your-org/mysql-analyzer-mcp/issues"

[project.scripts]
mysql-analyzer-mcp = "mysql_analyzer_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

# Black code formatting
[tool.black]
line-length = 99
target-version = ["py310"]
include = '\.pyi?$'

# Ruff linting and formatting
[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "Q"]
ignore = ["E203", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["mysql_analyzer_mcp"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# MyPy type checking
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
